refactor centralizeImage
